Crate memory_structs

This crate contains common types used for memory mapping.

Structs

AggregatedSectionMemoryBounds
The address bounds and flags of the initial kernel sections that need mapping.

Frame
A Frame is a chunk of physical memory, similar to how a Page is a chunk of virtual memory.

FrameRange
A range of Frames that are contiguous in physical memory.

Page
A virtual memory page, which contains the index of the page

PageRange
A range of Pages that are contiguous in virtual memory.

PhysicalAddress
A physical memory address, which is a usize under the hood.

PhysicalMemoryArea
An area of physical memory.

SectionMemoryBounds
The address bounds and mapping flags of a section's memory region.

VirtualAddress
A virtual memory address, which is a usize under the hood.