Struct memory_structs::PhysicalAddress

#[repr(transparent)]
pub struct PhysicalAddress(_);

A physical memory address, which is a usize under the hood.

Methods

impl PhysicalAddress

pub fn new(phys_addr: usize) -> Result<PhysicalAddress, &'static str>

Creates a new PhysicalAddress, checking that the bits (64:52] are 0.

pub fn new_canonical(phys_addr: usize) -> PhysicalAddress

Creates a new PhysicalAddress that is guaranteed to be canonical by forcing the upper bits (64:52] to be 0.

pub fn value(&self) -> usize

Returns the underlying usize value for this PhysicalAddress.

pub const fn zero() -> PhysicalAddress

Creates a PhysicalAddress with the value 0.

pub fn frame_offset(&self) -> usize

Returns the offset that this PhysicalAddress specifies into its containing memory Frame.

For example, if the PAGE_SIZE is 4KiB, then this will return the least significant 12 bits (12:0] of this PhysicalAddress.