Crate memory_structs

This crate contains common types used for memory mapping.

Structs

AggregatedSectionMemoryBounds
The address bounds and flags of the initial kernel sections that need mapping.

EntryFlags
Page table entry flags on the x86_64 architecture.

Frame
A Frame is a chunk of physical memory aligned to a PAGE_SIZE boundary.

FrameRange
A range of Frames that are contiguous in physical memory.

Page
A Page is a chunk of virtual memory aligned to a PAGE_SIZE boundary.

PageRange
A range of Pages that are contiguous in virtual memory.

PhysicalAddress
A physical memory address, which is a usize under the hood.

SectionMemoryBounds
The address bounds and mapping flags of a section’s memory region.

VirtualAddress
A virtual memory address, which is a usize under the hood.