Crate memory_structs

This crate contains basic types used for memory management.

The types of interest are divided into three categories:

addresses: VirtualAddress and PhysicalAddress.

“chunk” types: Page and Frame.

ranges of chunks: PageRange and FrameRange.

Structs

Frame

A Frame is a chunk of physical memory aligned to a page boundary (default 4KiB) given by the P parameter.

FrameRange

A range of Frames that are contiguous in physical memory.

Page

A Page is a chunk of virtual memory aligned to a page boundary (default 4KiB) given by the P parameter.

Page1G

Marker struct used to indicate a page size of 1GiB.

Page2M

Marker struct used to indicate a page size of 2MiB.

Page4K

Marker struct used to indicate the default page size of 4KiB.

PageRange

A range of Pages that are contiguous in virtual memory.

PhysicalAddress

A physical memory address, which is a usize under the hood.

VirtualAddress

A virtual memory address, which is a usize under the hood.

Enums

MemChunkSize

Enum used to indicate the size of a page or frame.

MemoryState

The possible states that a range of exclusively-owned pages or frames can be in.

Traits

PageSize

Trait that represents the size of a page or frame, i.e., for normal or huge pages.