Crate memory_structs

This crate contains basic types used for memory management.

The types of interest are divided into three categories:

addresses: VirtualAddress and PhysicalAddress.

“chunk” types: Page and Frame.

ranges of chunks: PageRange and FrameRange.

Structs

CopyableFrameRange

A FrameRange that implements Copy

CopyablePageRange

A PageRange that implements Copy

Frame

A Frame is a chunk of physical memory aligned to a [PAGE_SIZE] boundary.

FrameRange

A range of Frames that are contiguous in physical memory.

Page

A Page is a chunk of virtual memory aligned to a [PAGE_SIZE] boundary.

PageRange

A range of Pages that are contiguous in virtual memory.

PhysicalAddress

A physical memory address, which is a usize under the hood.

VirtualAddress

A virtual memory address, which is a usize under the hood.