Struct memory_structs::FrameRange

pub struct FrameRange(_);

A range of Frames that are contiguous in physical memory.

Implementations

impl FrameRange

pub const fn new(start: Frame, end: Frame) -> FrameRange

Creates a new range of Frames that spans from start to end, both inclusive bounds.

pub const fn empty() -> FrameRange

Creates a FrameRange that will always yield None when iterated.

pub fn from_phys_addr(
starting_addr: PhysicalAddress,
size_in_bytes: usize
) -> FrameRange

A convenience method for creating a new FrameRange that spans all Frames from the given PhysicalAddress to an end bound based on the given size.

pub const fn start_address(&self) -> PhysicalAddress

Returns the PhysicalAddress of the starting Frame in this FrameRange.

pub const fn size_in_frames(&self) -> usize

Returns the number of Frames covered by this iterator.

Use this instead of Iterator::count() method. This is instant, because it doesn’t need to iterate over each entry, unlike normal iterators.

pub const fn size_in_bytes(&self) -> usize

Returns the size of this range in number of bytes.

pub fn contains_address(&self, addr: PhysicalAddress) -> bool

Returns true if this FrameRange contains the given PhysicalAddress.

pub fn offset_of_address(&self, addr: PhysicalAddress) -> Option<usize>

Returns the offset of the given PhysicalAddress within this FrameRange, i.e., addr - self.start_address().

If the given addr is not covered by this range of Frames, this returns None.

Examples

If the range covers addresses 0x2000 to 0x4000, then offset_of_address(0x3500) would return Some(0x1500).

pub fn address_at_offset(&self, offset: usize) -> Option<PhysicalAddress>

Returns the PhysicalAddress at the given offset into this FrameRangewithin this FrameRange, i.e., addr - self.start_address().

If the given offset is not within this range of Frames, this returns None.

Examples

If the range covers addresses 0x2000 to 0x4000, then address_at_offset(0x1500) would return Some(0x3500).

pub fn to_extended(&self, to_include: Frame) -> FrameRange

Returns a new separate FrameRange that is extended to include the given Frame.

pub fn overlap(&self, other: &FrameRange) -> Option<FrameRange>

Returns an inclusive FrameRange representing the Frames that overlap across this FrameRange and the given other FrameRange.

If there is no overlap between the two ranges, None is returned.